+++ test/smokes/coffeelint/expectations.rb
@@ -1,6 +1,6 @@
 s = Runners::Testing::Smoke
 
-default_version = "4.1.2"
+default_version = "4.1.3"
 
 s.add_test(
   "success",