+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.42.1...HEAD)
 
+- **remark-lint** Add file path to error message [#2067](https://github.com/sider/runners/pull/2067)
+
 ## 0.42.1
 
 [Full diff](https://github.com/sider/runners/compare/0.42.0...0.42.1)